Move everything that relates to dummy app to a demo-app package
In the addon package, we keep only the dependencies and what relates to lint and release.
The rest is moved to a new demo-app private package that consumes the addon in its devDeps.
Root commands and CI instructions are adjusted to fit the new structure.

Note that the dependenciesMeta field in the demo app relates to an issue when ember/test-helper
is used with ember-try: https://github.com/emberjs/ember-test-helpers/issues/1489